+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Bug fixes:
   + Fixed compilation issues for nullable Int32 struct fields in JNI.
   + Fixed compilation issues for nullable unsigned integer struct fields in JNI.
+  + Fixed generated equality operator implementation for nullable struct fields in C++.
   + Fixed leaking JNI local references.
 
 ## 3.20.2